requirements.txt
yt-dlp
inaspeechsegmenter
shazamapi
pyyaml
filelock
jupyter